requirements.txt
flask
angel-one-smartapi
